+++ src/commonMain/kotlin/csense/kotlin/annotations/numbers/DoubleLimit.kt
@@ -1,5 +1,8 @@
+@file:Suppress("unused")
+
 package csense.kotlin.annotations.numbers
 
+//region Double limit
 @Target(
     AnnotationTarget.FUNCTION,
     AnnotationTarget.PROPERTY_GETTER,
@@ -15,3 +18,40 @@ annotation class DoubleLimit(
     val from: Double,
     val to: Double
 )
+//endregion
+
+//region Double limit to max
+@Target(
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY_GETTER,
+    AnnotationTarget.PROPERTY_SETTER,
+    AnnotationTarget.VALUE_PARAMETER,
+    AnnotationTarget.FIELD,
+    AnnotationTarget.LOCAL_VARIABLE,
+    AnnotationTarget.ANNOTATION_CLASS
+)
+
+@Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
+@MustBeDocumented
+annotation class DoubleLimitToMax(
+    val from: Double
+)
+//endregion
+
+//region Double limit from min
+@Target(
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY_GETTER,
+    AnnotationTarget.PROPERTY_SETTER,
+    AnnotationTarget.VALUE_PARAMETER,
+    AnnotationTarget.FIELD,
+    AnnotationTarget.LOCAL_VARIABLE,
+    AnnotationTarget.ANNOTATION_CLASS
+)
+
+@Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
+@MustBeDocumented
+annotation class DoubleLimitFromMin(
+    val from: Double
+)
+//endregion