+++ src/commonMain/kotlin/csense/kotlin/annotations/numbers/FloatLimit.kt
@@ -1,5 +1,8 @@
+@file:Suppress("unused")
+
 package csense.kotlin.annotations.numbers
 
+//region float limit
 @Target(
     AnnotationTarget.FUNCTION,
     AnnotationTarget.PROPERTY_GETTER,
@@ -16,3 +19,40 @@ annotation class FloatLimit(
     val from: Float,
     val to: Float
 )
+//endregion
+
+//region float limit to max
+@Target(
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY_GETTER,
+    AnnotationTarget.PROPERTY_SETTER,
+    AnnotationTarget.VALUE_PARAMETER,
+    AnnotationTarget.FIELD,
+    AnnotationTarget.LOCAL_VARIABLE,
+    AnnotationTarget.ANNOTATION_CLASS
+)
+
+@Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
+@MustBeDocumented
+annotation class FloatLimitToMax(
+    val from: Float
+)
+//endregion
+
+//region float limit from min
+@Target(
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY_GETTER,
+    AnnotationTarget.PROPERTY_SETTER,
+    AnnotationTarget.VALUE_PARAMETER,
+    AnnotationTarget.FIELD,
+    AnnotationTarget.LOCAL_VARIABLE,
+    AnnotationTarget.ANNOTATION_CLASS
+)
+
+@Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
+@MustBeDocumented
+annotation class FloatLimitFromMin(
+    val from: Float
+)
+//endregion