+++ src/commonMain/kotlin/csense/kotlin/annotations/threading/InUIContext.kt
@@ -1,10 +1,12 @@
 @file:Suppress("unused")
+
 package csense.kotlin.annotations.threading
 
 @Retention(AnnotationRetention.BINARY)
 @Target(
     AnnotationTarget.CLASS,
     AnnotationTarget.FILE,
-    AnnotationTarget.TYPE)
+    AnnotationTarget.TYPE
+)
 @MustBeDocumented
 annotation class InUiContext